test: merge Mesh-API dataclass/equals/repr into test_api_mesh
